Move message parsing libs/MessageParser.js. Move html renderer to libs/renderers/Html.js.
